+++ src/Player/AABB.cpp
@@ -1,6 +1,6 @@
 #include "AABB.h"
 
-AABB::AABB(glm::vec3 min, glm::vec3 max)
+AABB::AABB(glm::dvec3 min, glm::dvec3 max)
 	: m_Min{ min },
 	m_Max{ max }
 {
@@ -10,12 +10,12 @@ AABB::AABB()
 {
 }
 
-bool operator< (const glm::vec3& v1, const glm::vec3& v2)
+bool operator< (const glm::dvec3& v1, const glm::dvec3& v2)
 {
 	return (v1.x < v2.x) && (v1.y < v2.y) && (v1.z < v2.z);
 }
 
-bool operator> (const glm::vec3& v1, const glm::vec3& v2)
+bool operator> (const glm::dvec3& v1, const glm::dvec3& v2)
 {
 	return (v1.x > v2.x) && (v1.y > v2.y) && (v1.z > v2.z);
 }
@@ -25,22 +25,22 @@ bool AABB::intersects(const AABB& other)
 	return (m_Min < other.max() && m_Max > other.min());
 }
 
-glm::vec3 AABB::min() const
+glm::dvec3 AABB::min() const
 {
 	return m_Min;
 }
 
-glm::vec3 AABB::max() const
+glm::dvec3 AABB::max() const
 {
 	return m_Max;
 }
 
-void AABB::min(glm::vec3 v)
+void AABB::min(glm::dvec3 v)
 {
 	m_Min = v;
 }
 
-void AABB::max(glm::vec3 v)
+void AABB::max(glm::dvec3 v)
 {
 	m_Max = v;
 }
